added sorting to data table

+++ src/app/admin/admin-products/product-create/product-create.component.ts
@@ -21,7 +21,6 @@ export class ProductCreateComponent implements OnInit {
               private productService: ProductService) {}
 
   ngOnInit() {
-    console.log("inside create");
     this.categoryService.getAllCategory()
     .subscribe(categories => {
       this.categories = categories;
@@ -30,7 +29,6 @@ export class ProductCreateComponent implements OnInit {
     this.id = this.route.snapshot.paramMap.get('id');
     if(this.id) {
       this.product = this.productService.getProductById(this.id);
-      console.log(this.product);
     }
 
   }
@@ -41,92 +39,22 @@ export class ProductCreateComponent implements OnInit {
        return c.categoryId == newProduct.category; 
     })
     newProduct.categoryName = selectedCategory.categoryName;
-    this.productService.addProduct(newProduct);
+    if(this.id){
+      this.productService.editProduct(newProduct, this.id);
+    }else {
+      this.productService.addProduct(newProduct);
+    }
+    
   }
 
   onDelete() {
+    if (!confirm('Are you sure you want to delete this product?')) return;
+
     this.productService.deleteProduct(this.id);
   }
 
-
-
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
-  // id: string;
-  // editMode: boolean = false;
-  // productForm: FormGroup;
-
-  // constructor(private route: ActivatedRoute,
-  //             private router: Router) {}
-
-  // ngOnInit() {
-  //   if(this.route.snapshot.params['id']){
-  //     this.editMode = true;
-  //     this.id = this.route.snapshot.params['id'];
-  //   }
-  //   this.initForm();
-  // }
-
-  // private initForm(){
-  //   let productCategory = '';
-  //   let productTitle = '';
-  //   let productImageUrl = '';
-  //   let productPrice = 0;
-
-  //   if(this.editMode){
-  //     const product = this.productService.getProduct(this.id);
-  //     productCategory = product.category;
-  //     productTitle = product.title;
-  //     productImageUrl = product.imageUrl;
-  //     productPrice = product.price;
-  //   }
-
-  //   this.productForm = new FormGroup({
-  //     'category': new FormControl(productCategory, Validators.required),
-  //     'imageUrl': new FormControl(productImageUrl, Validators.required),
-  //     'title': new FormControl(productTitle, Validators.required),
-  //     'price': new FormControl(productPrice, Validators.required),
-  //   });
-  // }
-
-  // onSubmit(){
-  //   const title = this.productForm.value.title;
-  //   const category = this.productForm.value.category;
-  //   const price = this.productForm.value.price
-  //   const imageUrl = this.productForm.value.imageUrl
-  //   if(this.editMode){
-  //     const newProduct = new Product(this.id, category, title, imageUrl, price, false);
-  //     this.productService.updateProduct(this.id, newProduct);
-  //   } else {
-  //     const newProduct = new Product(null, category, title, imageUrl, price, false);
-  //     this.productService.addProduct(newProduct);
-  //   }
-  //   this.router.navigate(['../'], {relativeTo: this.route});
-  // }
-
-  // onCancel(){
-  //   this.router.navigate(['../'], {relativeTo: this.route});
-  // }
+  onCancel() {
+    this.router.navigate(['../'],{relativeTo: this.route});
+  }
 
 }
